Move monster card test to detox
I had to change the accessibilityLabel for the view containing the cards
because it's defaulted by concatenating the labels from the child
elements. So when there was only one card, the container would have the
same accessibilityLabel as the card (odd that this only became apparent
in detox, but might have something to do with how RN testing library
works vs actual React Native).