Fix build
Had to remove the animations because for some reason it was possible to
have cards overlap when using the animated layout stuff. I think it had
to do with the shared value thing too.

The issue was, when you'd change the room, if a card was going to show
up in the new room too, it wouldn't pick up on a new position.

I'm thinking this might be resolved if I use the normal React Native
animated stuff, instead of Reanimated, but I don't want to bother with
that at the moment.

The test that caught this issue was "can be done after not running a
round" in the "running" section.